+++ src/bluez/adapter/mod.rs
@@ -17,18 +17,18 @@ mod peripheral;
 use libc;
 use nom;
 use bytes::{BytesMut, BufMut};
+use dashmap::DashMap;
 
 use std::{
     self,
     ffi::CStr,
-    collections::{HashSet, HashMap},
-    sync::{Arc, Mutex},
-    sync::atomic::{AtomicBool, Ordering},
+    collections::HashSet,
+    sync::{Arc, mpsc::Receiver, atomic::{AtomicBool, Ordering}},
     thread,
 };
 
 use crate::Result;
-use crate::api::{CentralEvent, BDAddr, Central, EventHandler};
+use crate::api::{CentralEvent, BDAddr, Central, AdapterManager};
 
 use crate::bluez::{
     util::handle_error,
@@ -180,7 +180,7 @@ impl AdapterState {
 }
 
 /// The [`Central`](../../api/trait.Central.html) implementation for BlueZ.
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct ConnectedAdapter {
     pub adapter: Adapter,
     adapter_fd: i32,
@@ -188,9 +188,8 @@ pub struct ConnectedAdapter {
     pub scan_enabled: Arc<AtomicBool>,
     pub active: Arc<AtomicBool>,
     pub filter_duplicates: Arc<AtomicBool>,
-    peripherals: Arc<Mutex<HashMap<BDAddr, Peripheral>>>,
-    handle_map: Arc<Mutex<HashMap<u16, BDAddr>>>,
-    event_handlers: Arc<Mutex<Vec<EventHandler>>>,
+    handle_map: Arc<DashMap<u16, BDAddr>>,
+    manager: AdapterManager<Peripheral>
 }
 
 impl ConnectedAdapter {
@@ -219,9 +218,8 @@ impl ConnectedAdapter {
             filter_duplicates: Arc::new(AtomicBool::new(false)),
             should_stop,
             scan_enabled: Arc::new(AtomicBool::new(false)),
-            event_handlers: Arc::new(Mutex::new(vec![])),
-            peripherals: Arc::new(Mutex::new(HashMap::new())),
-            handle_map: Arc::new(Mutex::new(HashMap::new())),
+            handle_map: Arc::new(DashMap::new()),
+            manager: AdapterManager::new(),
         };
 
         connected.add_raw_socket_reader(adapter_fd);
@@ -300,11 +298,7 @@ impl ConnectedAdapter {
 
     fn emit(&self, event: CentralEvent) {
         debug!("emitted {:?}", event);
-        let handlers = self.event_handlers.clone();
-        let vec = handlers.lock().unwrap();
-        for handler in (*vec).iter() {
-            handler(event.clone());
-        }
+        self.manager.emit(event);
     }
 
     fn handle(&self, message: hci::Message) {
@@ -312,24 +306,16 @@ impl ConnectedAdapter {
 
         match message {
             hci::Message::LEAdvertisingReport(info) => {
-                let mut new = false;
                 let address = info.bdaddr.clone();
-
-                {
-                    let mut peripherals = self.peripherals.lock().unwrap();
-                    let peripheral = peripherals.entry(info.bdaddr)
-                        .or_insert_with(|| {
-                            new = true;
-                            Peripheral::new(self.clone(), info.bdaddr)
-                        });
-
-
-                    peripheral.handle_device_message(&hci::Message::LEAdvertisingReport(info));
-                }
-
-                if new {
+                let info_clone = info.clone();
+                let peripheral = Peripheral::new(self.clone(), info.bdaddr);
+                peripheral.handle_device_message(&hci::Message::LEAdvertisingReport(info));
+                if !self.manager.has_peripheral(&address) {
+                    warn!("{:?}", info_clone);
+                    self.manager.add_peripheral(address, peripheral);                    
                     self.emit(CentralEvent::DeviceDiscovered(address.clone()))
                 } else {
+                    self.manager.update_peripheral(address, peripheral);
                     self.emit(CentralEvent::DeviceUpdated(address.clone()))
                 }
             }
@@ -345,8 +331,7 @@ impl ConnectedAdapter {
                     None => warn!("Got connection for unknown device {}", info.bdaddr)
                 }
 
-                let mut handles = self.handle_map.lock().unwrap();
-                handles.insert(handle, address);
+                self.handle_map.insert(handle, address);
 
                 self.emit(CentralEvent::DeviceConnected(address));
             }
@@ -355,22 +340,21 @@ impl ConnectedAdapter {
 
                 // TODO this is a bit risky from a deadlock perspective (note mutexes are not
                 // reentrant in rust!)
-                let peripherals = self.peripherals.lock().unwrap();
+                let peripherals = self.manager.peripherals();
 
-                for peripheral in peripherals.values() {
+                for peripheral in peripherals {
                     // we don't know the handler => device mapping, so send to all and let them filter
                     peripheral.handle_device_message(&message);
                 }
             },
             hci::Message::DisconnectComplete { handle, .. } => {
-                let mut handles = self.handle_map.lock().unwrap();
-                match handles.remove(&handle) {
+                match self.handle_map.remove(&handle) {
                     Some(addr) => {
-                        match self.peripheral(addr) {
+                        match self.peripheral(addr.1) {
                             Some(peripheral) => peripheral.handle_device_message(&message),
-                            None => warn!("got disconnect for unknown device {}", addr),
+                            None => warn!("got disconnect for unknown device {}", addr.1),
                         };
-                        self.emit(CentralEvent::DeviceDisconnected(addr));
+                        self.emit(CentralEvent::DeviceDisconnected(addr.1));
                     }
                     None => {
                         warn!("got disconnect for unknown handle {}", handle);
@@ -415,9 +399,8 @@ impl ConnectedAdapter {
 }
 
 impl Central<Peripheral> for ConnectedAdapter {
-    fn on_event(&self, handler: EventHandler) {
-        let list = self.event_handlers.clone();
-        list.lock().unwrap().push(handler);
+    fn event_receiver(&self) -> Option<Receiver<CentralEvent>> {
+        self.manager.event_receiver()
     }
 
     fn active(&self, enabled: bool) {
@@ -438,13 +421,11 @@ impl Central<Peripheral> for ConnectedAdapter {
     }
 
     fn peripherals(&self) -> Vec<Peripheral> {
-        let l = self.peripherals.lock().unwrap();
-        l.values().map(|p| p.clone()).collect()
+        self.manager.peripherals()
     }
 
     fn peripheral(&self, address: BDAddr) -> Option<Peripheral> {
-        let l = self.peripherals.lock().unwrap();
-        l.get(&address).map(|p| p.clone())
+        self.manager.peripheral(address)
     }
 }